third_party/blink/web_tests/fast/selectors/placeholder-shown-sibling-style-update-expected.txt

Test style update of the :placeholder-shown pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial state is without placehoder.
PASS getComputedStyle(document.getElementById("target1")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("target2")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("target3")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("target4")).backgroundColor is "rgb(255, 255, 255)"
Adding a placeholder, the targets should match the style.
PASS getComputedStyle(document.getElementById("target1")).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.getElementById("target2")).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.getElementById("target3")).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.getElementById("target4")).backgroundColor is "rgb(1, 2, 3)"
PASS successfullyParsed is true

TEST COMPLETE

 
